requirements.txt
# Database dependencies
psycopg2-binary==2.9.11

# HTTP requests
requests==2.32.5

# RabbitMQ message broker
pika==1.3.2

# Environment variables management
python-dotenv==1.2.1

# Additional dependencies
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3

# Optional: Development dependencies (uncomment if needed)
# pytest==7.4.3
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
